requirements.txt
beautifulsoup4==4.12.2
celery==5.3.1
flake8==6.0.0
Flask==2.3.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.17
redis==4.6.0
